requirements.txt
flask
firebase_admin
python-dotenv
pyrebase4
bootstrap-flask
flask-WTF